+++ server/models/Users.js
@@ -1,0 +1,80 @@
+const { DataTypes } = require("sequelize");
+const bcrypt = require("bcryptjs");
+
+module.exports = (sequelize) => {
+  const Users = sequelize.define(
+    "Users",
+    {
+      id: {
+        type: DataTypes.INTEGER,
+        autoIncrement: true,
+        primaryKey: true,
+      },
+      name: {
+        type: DataTypes.STRING(100),
+        allowNull: false,
+      },
+      email: {
+        type: DataTypes.STRING(255),
+        allowNull: false,
+        unique: true,
+        validate: {
+          isEmail: true,
+        },
+      },
+      phoneNumber: {
+        type: DataTypes.STRING(30),
+        allowNull: true,
+      },
+      password: {
+        type: DataTypes.STRING(255),
+        allowNull: false,
+      },
+      userType: {
+        type: DataTypes.ENUM("customer", "admin", "vip"),
+        allowNull: false,
+        defaultValue: "customer",
+      },
+      profilePic: {
+        type: DataTypes.STRING(255),
+        allowNull: true,
+      },
+      verified: {
+        type: DataTypes.BOOLEAN,
+        allowNull: false,
+        defaultValue: false,
+      },
+      verificationCode: {
+        type: DataTypes.STRING(20),
+        allowNull: true,
+      },
+      passwordResetToken: {
+        type: DataTypes.STRING(255),
+        allowNull: true,
+      },
+      passwordResetExpires: {
+        type: DataTypes.DATE,
+        allowNull: true,
+      },
+    },
+    {
+      tableName: "Users",
+      timestamps: true,
+      paranoid: true,
+    }
+  );
+
+  Users.beforeCreate(async (user) => {
+    if (user.password) {
+      user.password = await bcrypt.hash(user.password, 10);
+    }
+  });
+
+  Users.beforeUpdate(async (user) => {
+    if (user.changed("password") && user.password) {
+      user.password = await bcrypt.hash(user.password, 10);
+    }
+  });
+
+  return Users;
+};